move PostTweetsProvider under TweetProvider namespace
